deprecated usage of class specific validation algorithms for speed, heading, course, turn, and position in favor of static methods in AISMessage, incremented version number to 2.1.7

+++ src/main/java/jais/messages/AISMessage.java
@@ -99,7 +99,7 @@ public interface AISMessage {
      * @param courseOverGround the value to evaluate
      * @return a boolean indicating whether the course information is valid or not
      */
-    public static boolean isValidCourse(int courseOverGround) {
+    public static boolean isValidCourse(float courseOverGround) {
         return courseOverGround < 3600;
     }
